+++ byconServices/intervalFrequencies.py
@@ -2,9 +2,9 @@ from bycon import BYC_PARS, BeaconErrorResponse
 from byconServiceLibs import ByconBundler, GenomeBins, ByconServiceResponse, PGXfreq
 
 """
-* https://progenetix.org/services/intervalFrequencies/?datasetIds=progenetix&filters=NCIT:C7376,PMID:22824167,pgx:icdom-85003
+* https://progenetix.org/services/intervalFrequencies/?datasetIds=progenetix&filters=NCIT:C7376,pubmed:22824167,pgx:icdom-85003
 * https://progenetix.org/services/intervalFrequencies/?datasetIds=progenetix&id=pgx:cohort-TCGAcancers
-* http://progenetix.test/services/intervalFrequencies/?datasetIds=progenetix&output=pgxmatrix&filters=NCIT:C7376,PMID:22824167
+* http://progenetix.test/services/intervalFrequencies/?datasetIds=progenetix&output=pgxmatrix&filters=NCIT:C7376,pubmed:22824167
 """
 
 def intervalFrequencies():