+++ updev.sh
@@ -14,4 +14,4 @@ rm -rf ./build
 rm -rf ./dist
 rm -rf ./bycon.egg-info
 # rsync -avh --exclude=logs --exclude=.git ./ ~/Github/bycon/
-# git tag v2.1.0  & git push --tags
+# git tag v2.2.6  & git push --tags